Imported tasks now imported as a tree structure
To avoid having to explicitly import each task in an external YML or JavaScript file, we want to provide a means to allow us to import all tasks in an imported file. To avoid a clash of task names across imported files, we now create a tree structure of tasks - namespaced after the imported file/package name. Thus we can have tasks like bastion.create or phoenix.create.