Move dereference object into shell command
The generic state.get tries to deference the object passed to it. One of the issues is that - on occasions - we need to deference the values inside an object.  We did add the code to dereference the values inside the state get function, but there is an issue here. First, it de-refs any object passed to it, and that includes commands in a sequence object. This is premature, some of the commands in the object cannot be de-referenced until the previous command is run (becuase the previous command may set variables). 

However, the environment variables is an object which does need to be de-referenced, but we have no way to tell inside the generic de-ref code which object to de-ref and which not to.

Solution was to explicity de-ref the environments vars inside shell command and leave the generic code to just return the object (which is the default in the majority of the cases.)